requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pydantic==2.10.2
alembic==1.14.0
psycopg2==2.9.10
ecdsa==0.19.0
cryptography==44.0.0
PyQt6==6.7.1
